+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 
 ## [Unreleased]
 - (VANAGON-184) Remove support for Fedora 32 (x86-64)
+- (VANAGON-196) Remove support for Ubuntu 16.04 (x86 and x86-64)
 
 ### Changed